Add ref guard for PersistedLazy
This is more challening than the ref guard on Persisted because a mutable access to PersistedLazy does not necessarily mean the persisted value is going to change, and we want to avoid unnecessary persistence calls. The solution here is to cache the most recent persisted value, and only do the write if it's changed. This means the only extra work we have to do on each mutable access is derive the persisted value (which is meant to be cheap), then compare it to the existing one. It is *some* overhead, but at least it doesn't involve I/O or anything fallible.