Rename methods on PersistedContainer
- `get_persisted` -> `get_to_persist`
- `set_persisted` -> `restore_persisted`

I struggled to remember which method is which at a glance, because the data is bi-directional (`set_persisted` sounds like it sets the value in the store but it's the opposite). These names should be more clear. I don't love them but it's the best I could come up with.